Moving logic to delayed.js
